Add local modal dialog after MUI example. Remove fi-sci dependency.
