+++ gui/package.json
@@ -23,10 +23,9 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@fi-sci/misc": "^0.0.8",
-    "@fi-sci/modal-window": "^0.0.2",
     "@geoffcox/react-splitter": "^2.1.2",
     "@monaco-editor/react": "^4.6.0",
-    "@mui/icons-material": "^5.15.17",
+    "@mui/icons-material": "^5.16.5",
     "@mui/material": "^5.15.17",
     "@octokit/rest": "^21.0.0",
     "@vercel/analytics": "^1.3.1",
@@ -65,5 +64,6 @@
     "vite": "^5.2.12",
     "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.6.0"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha1.ac34549e6aa8e7ead463a7407e1c7390f61a6610"
 }
