+++ gui/src/app/App.tsx
@@ -1,17 +1,23 @@
-import { BrowserRouter } from "react-router-dom";
+import { createTheme } from "@mui/material";
+import { ThemeProvider } from "@mui/system";
 import ProjectContextProvider from "@SpCore/ProjectContextProvider";
 import HomePage from "@SpPages/HomePage";
 import { Analytics } from "@vercel/analytics/react";
+import { BrowserRouter } from "react-router-dom";
+
+const theme = createTheme();
 
 function App() {
   return (
     <BrowserRouter>
-      <div className="MainWindow">
-        <ProjectContextProvider>
-          <HomePage />
-        </ProjectContextProvider>
-        <Analytics />
-      </div>
+      <ThemeProvider theme={theme}>
+        <div className="MainWindow">
+          <ProjectContextProvider>
+            <HomePage />
+          </ProjectContextProvider>
+          <Analytics />
+        </div>
+      </ThemeProvider>
     </BrowserRouter>
   );
 }
