+++ gui/src/app/CompilationServerConnectionControl/CompilationServerConnectionControl.tsx
@@ -1,16 +1,19 @@
 import { SmallIconButton } from "@fi-sci/misc";
-import { default as ModalWindow, useModalWindow } from "@fi-sci/modal-window";
 import { Cancel, Check } from "@mui/icons-material";
-import Button from "@mui/material/Button";
 import Link from "@mui/material/Link";
+
+import CloseableDialog, {
+  useDialogControls,
+} from "@SpComponents/CloseableDialog";
 import { FunctionComponent, useCallback, useEffect, useState } from "react";
+import ConfigureCompilationServerDialog from "./ConfigureCompilationServerDialog";
+
+export const publicUrl = "https://trom-stan-wasm-server.magland.org";
+export const localUrl = "http://localhost:8083";
 
 type CompilationServerConnectionControlProps = {
   // none
 };
-
-const publicUrl = "https://trom-stan-wasm-server.magland.org";
-const localUrl = "http://localhost:8083";
 
 const CompilationServerConnectionControl: FunctionComponent<
   CompilationServerConnectionControlProps
@@ -26,8 +29,8 @@ const CompilationServerConnectionControl: FunctionComponent<
   const {
     handleOpen: openDialog,
     handleClose: closeDialog,
-    visible: dialogVisible,
-  } = useModalWindow();
+    open,
+  } = useDialogControls();
 
   const handleRetry = useCallback(() => {
     retryConnection();
@@ -60,7 +63,12 @@ const CompilationServerConnectionControl: FunctionComponent<
         </Link>
         &nbsp;&nbsp;
       </span>
-      <ModalWindow visible={dialogVisible} onClose={closeDialog}>
+      <CloseableDialog
+        title="Select a compilation server"
+        id="compilationDialog"
+        open={open}
+        handleClose={closeDialog}
+      >
         <ConfigureCompilationServerDialog
           stanWasmServerUrl={stanWasmServerUrl}
           setStanWasmServerUrl={setStanWasmServerUrl}
@@ -68,7 +76,7 @@ const CompilationServerConnectionControl: FunctionComponent<
           closeDialog={closeDialog}
           onRetry={handleRetry}
         />
-      </ModalWindow>
+      </CloseableDialog>
     </span>
   );
 };
@@ -94,125 +102,6 @@ const useIsConnected = (stanWasmServerUrl: string) => {
     })();
   }, [probeUrl, retryCode]);
   return { isConnected, retryConnection };
-};
-
-type ConfigureCompilationServerDialogProps = {
-  stanWasmServerUrl: string;
-  setStanWasmServerUrl: (url: string) => void;
-  isConnected: boolean;
-  closeDialog: () => void;
-  onRetry: () => void;
-};
-
-const ConfigureCompilationServerDialog: FunctionComponent<
-  ConfigureCompilationServerDialogProps
-> = ({
-  stanWasmServerUrl,
-  setStanWasmServerUrl,
-  isConnected,
-  closeDialog,
-  onRetry,
-}) => {
-  const [choice, setChoice] = useState<"public" | "local" | "custom">("custom");
-  useEffect(() => {
-    if (stanWasmServerUrl === publicUrl) setChoice("public");
-    else if (stanWasmServerUrl === localUrl) setChoice("local");
-    else setChoice("custom");
-  }, [stanWasmServerUrl]);
-  return (
-    <div>
-      <h3>Select a compilation server</h3>
-      <p>
-        While the sampling is performed locally in the browser, a compilation
-        server is required to compile the Stan programs.
-      </p>
-      <hr />
-      <div>
-        {isConnected ? (
-          <span className="connected">Connected</span>
-        ) : (
-          <span className="disconnected">Not connected</span>
-        )}
-        &nbsp;
-        <Link onClick={onRetry} component="button" underline="none">
-          retry
-        </Link>
-      </div>
-      <hr />
-      <div>
-        <input
-          type="radio"
-          id="public"
-          name="server"
-          value="public"
-          checked={choice === "public"}
-          onChange={() => {
-            if (choice === "custom") setChoice("public");
-            setStanWasmServerUrl(publicUrl);
-          }}
-        />
-        <label htmlFor="public">Public server</label>
-        <br />
-
-        <input
-          type="radio"
-          id="local"
-          name="server"
-          value="local"
-          checked={choice === "local"}
-          onChange={() => {
-            if (choice === "custom") setChoice("local");
-            setStanWasmServerUrl(localUrl);
-          }}
-        />
-        <label htmlFor="local">Local server</label>
-        <br />
-
-        <input
-          type="radio"
-          id="custom"
-          name="server"
-          value="custom"
-          checked={choice === "custom"}
-          onChange={() => setChoice("custom")}
-        />
-        <label htmlFor="custom">Custom server</label>
-        <br />
-
-        <input
-          // This one isn't honoring a class-based style for some reason
-          style={{ width: 500 }}
-          disabled={choice !== "custom"}
-          type="text"
-          value={stanWasmServerUrl}
-          onChange={(e) => setStanWasmServerUrl(e.target.value)}
-        />
-        <br />
-        <hr />
-        {choice === "local" && (
-          <div>
-            <p>To start a local compilation server:</p>
-            <div>
-              <pre>
-                docker run -p 8083:8080 -it magland/stan-wasm-server:latest
-              </pre>
-            </div>
-          </div>
-        )}
-        {choice === "public" && (
-          <div>
-            <p>
-              The public server is provided for convenience, but may not be as
-              reliable as a local server, depending on the current load and
-              availability.
-            </p>
-          </div>
-        )}
-        <hr />
-        <Button onClick={() => closeDialog()}>Close</Button>
-      </div>
-    </div>
-  );
 };
 
 export default CompilationServerConnectionControl;
