+++ gui/src/localStyles.css
@@ -13,6 +13,10 @@
   color: red;
 }
 
+.dialogWrapper {
+  padding: 10px;
+}
+
 .CompilationServerConnectionPane {
   font-size: 12px;
 }
